+++ back-end/src/main/kotlin/com/apexdevs/backend/web/controller/api/ApiWorkflowController.kt
@@ -6,21 +6,33 @@ package com.apexdevs.backend.web.controller.api
 
 import com.apexdevs.backend.ape.ApeRequestFactory
 import com.apexdevs.backend.ape.entity.workflow.Constraint
+import com.apexdevs.backend.ape.entity.workflow.Data
 import com.apexdevs.backend.ape.entity.workflow.InputData
 import com.apexdevs.backend.ape.entity.workflow.Ontology
 import com.apexdevs.backend.ape.entity.workflow.RunConfig
 import com.apexdevs.backend.ape.entity.workflow.TotalConfig
 import com.apexdevs.backend.ape.entity.workflow.WorkflowOutput
+import com.apexdevs.backend.ape.entity.workflow.constraintsFromJSON
+import com.apexdevs.backend.ape.entity.workflow.dataListFromJSON
+import com.apexdevs.backend.persistence.DomainOperation
+import com.apexdevs.backend.persistence.RunParametersOperation
 import com.apexdevs.backend.persistence.UserOperation
+import com.apexdevs.backend.persistence.database.entity.DomainAccess
+import com.apexdevs.backend.persistence.database.entity.DomainVerification
 import com.apexdevs.backend.persistence.database.entity.UserStatus
 import com.apexdevs.backend.persistence.exception.RunParametersExceedLimitsException
 import com.apexdevs.backend.persistence.exception.SynthesisFlagException
 import com.apexdevs.backend.persistence.filesystem.FileTypes
 import com.apexdevs.backend.persistence.filesystem.StorageService
+import com.apexdevs.backend.web.controller.entity.domain.DomainVerificationResult
+import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
+import com.fasterxml.jackson.module.kotlin.readValue
+import nl.uu.cs.ape.sat.models.enums.SynthesisFlag
 import org.json.JSONObject
 import org.springframework.core.io.Resource
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
+import org.springframework.security.access.AccessDeniedException
 import org.springframework.security.core.annotation.AuthenticationPrincipal
 import org.springframework.security.core.userdetails.User
 import org.springframework.web.bind.annotation.GetMapping
@@ -31,6 +43,7 @@ import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.ResponseStatus
 import org.springframework.web.bind.annotation.RestController
 import org.springframework.web.server.ResponseStatusException
+import java.io.FileNotFoundException
 import javax.servlet.http.HttpSession
 
 /**
@@ -40,13 +53,17 @@ import javax.servlet.http.HttpSession
  * @param apeRequestFactory factory that maintains the ApeRequest instances
  * @param storageService handles filesystem operations
  * @param userOperation operates on database user information
+ * @param domainOperation operates on database of domains
+ * @param runParametersOperation operates on database of run parameters
  */
 @RestController
 @RequestMapping("/api/workflow")
 class ApiWorkflowController(
     val apeRequestFactory: ApeRequestFactory,
     val storageService: StorageService,
-    val userOperation: UserOperation
+    val userOperation: UserOperation,
+    val domainOperation: DomainOperation,
+    val runParametersOperation: RunParametersOperation,
 ) {
     /**
      * @Return: workflow input and output data
@@ -162,9 +179,173 @@ class ApiWorkflowController(
     }
 
     /**
+     * Verify a domain can run without errors.
+     */
+    @ResponseStatus(HttpStatus.OK)
+    @GetMapping("/verify/ontology")
+    fun verifyOntology(@AuthenticationPrincipal user: User, session: HttpSession): DomainVerificationResult {
+        try {
+            val apeRequest = apeRequestFactory.getApeRequest(session.id)
+
+            val userResult = userOperation.getByEmail(user.username)
+            val hasAccess = domainOperation.hasUserAccess(apeRequest.domain, DomainAccess.ReadWrite, userResult.id)
+            if (!hasAccess)
+                throw AccessDeniedException("User does not have access to edit the domain")
+
+            val currentVerification = domainOperation.getVerification(apeRequest.domain.id)
+            val current = if (currentVerification.isPresent) {
+                currentVerification.get().asResult()
+            } else {
+                DomainVerificationResult()
+            }
+
+            val runConfig = RunConfig(
+                solutionMinLength = 1,
+                solutionMaxLength = 1,
+                maxSolutionsToReturn = 1,
+                maxDuration = 60,
+            )
+            storageService.storeConstraint(apeRequest.domain.id, FileTypes.Constraints, emptyList())
+            val result: DomainVerificationResult = try {
+                apeRequest.getWorkflows(runConfig)
+                DomainVerificationResult(true, current.useCaseSuccess, null)
+            } catch (exc: SynthesisFlagException) {
+                if (exc.flag == SynthesisFlag.UNSAT) {
+                    // No solutions found is considered correct
+                    DomainVerificationResult(true, current.useCaseSuccess, null)
+                } else
+                    DomainVerificationResult(false, current.useCaseSuccess, exc.message)
+            }
+            domainOperation.saveVerification(DomainVerification(apeRequest.domain.id, result))
+            return result
+        } catch (exc: Exception) {
+            when (exc) {
+                is AccessDeniedException ->
+                    throw ResponseStatusException(HttpStatus.FORBIDDEN, exc.message, exc)
+                else ->
+                    throw ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "", exc)
+            }
+        }
+    }
+
+    /**
+     * Verify a domain's use case configuration can run without errors.
+     */
+    @ResponseStatus(HttpStatus.OK)
+    @GetMapping("/verify/useCase")
+    fun verifyUseCase(@AuthenticationPrincipal user: User, session: HttpSession): DomainVerificationResult {
+        try {
+            val apeRequest = apeRequestFactory.getApeRequest(session.id)
+
+            val userResult = userOperation.getByEmail(user.username)
+            val hasAccess = domainOperation.hasUserAccess(apeRequest.domain, DomainAccess.ReadWrite, userResult.id)
+            if (!hasAccess)
+                throw AccessDeniedException("User does not have access to edit the domain")
+
+            val currentVerification = domainOperation.getVerification(apeRequest.domain.id)
+            val current = if (currentVerification.isPresent) {
+                currentVerification.get().asResult()
+            } else {
+                DomainVerificationResult()
+            }
+            val runParameters = runParametersOperation.getGlobalRunParameters()
+
+            try {
+                val mapper = jacksonObjectMapper()
+
+                val useCaseConfigJson = apeRequestFactory.fileService.loadFile(apeRequest.domain.id, FileTypes.UseCaseRunConfig)
+                val useCaseConstraintsJson = apeRequestFactory.fileService.loadFile(apeRequest.domain.id, FileTypes.UseCaseConstraints)
+                val useCaseConfig: Map<String, Any> = mapper.readValue(useCaseConfigJson)
+                val useCaseConstraints: Map<String, Any> = mapper.readValue(useCaseConstraintsJson)
+
+                val inputs: List<Data> =
+                    dataListFromJSON(useCaseConfig["inputs"] as ArrayList<LinkedHashMap<String, ArrayList<String>>>)
+                val outputs: List<Data> =
+                    dataListFromJSON(useCaseConfig["outputs"] as ArrayList<LinkedHashMap<String, ArrayList<String>>>)
+                val constraints: List<Constraint> =
+                    constraintsFromJSON(useCaseConstraints["constraints"] as ArrayList<LinkedHashMap<String, Any>>)
+                val solutionLengths: LinkedHashMap<String, Int> =
+                    useCaseConfig["solution_length"] as LinkedHashMap<String, Int>
+                val runConfig = RunConfig(
+                    solutionMinLength = solutionLengths["min"]!!,
+                    solutionMaxLength = solutionLengths["max"]!!,
+                    maxSolutionsToReturn = useCaseConfig["max_solutions"].toString().toInt(),
+                    maxDuration = useCaseConfig["timeout_sec"].toString().toInt(),
+                    inputs = inputs,
+                    outputs = outputs,
+                )
+
+                // Check if the use case configuration does not exceed the allowed run parameters
+                val result: DomainVerificationResult = if (runConfig.solutionMinLength > runParameters.minLength) {
+                    DomainVerificationResult(
+                        ontologySuccess = true,
+                        useCaseSuccess = false,
+                        "Use case solution minimum length exceeds the allowed minimum length " +
+                            "(Use case ${runConfig.solutionMinLength} vs allowed ${runParameters.minLength})"
+                    )
+                } else if (runConfig.solutionMaxLength > runParameters.maxLength) {
+                    DomainVerificationResult(
+                        ontologySuccess = true,
+                        useCaseSuccess = false,
+                        "Use case solution maximum length exceeds the allowed maximum length " +
+                            "(Use case ${runConfig.solutionMaxLength} vs allowed ${runParameters.maxLength})"
+                    )
+                } else if (runConfig.maxSolutionsToReturn > runParameters.solutions) {
+                    DomainVerificationResult(
+                        ontologySuccess = true,
+                        useCaseSuccess = false,
+                        "Use case max solutions exceeds the allowed max solutions " +
+                            "(Use case ${runConfig.maxSolutionsToReturn} vs allowed ${runParameters.solutions})"
+                    )
+                } else if (runConfig.maxDuration > runParameters.maxDuration) {
+                    DomainVerificationResult(
+                        ontologySuccess = true,
+                        useCaseSuccess = false,
+                        "Use case max duration exceeds the allowed max duration " +
+                            "(Use case ${runConfig.maxDuration} vs allowed ${runParameters.maxDuration})"
+                    )
+                } else { // Run the configuration and check requirements for verification
+                    storageService.storeConstraint(apeRequest.domain.id, FileTypes.Constraints, constraints)
+                    val workflows = apeRequest.getWorkflows(runConfig)
+                    if (workflows.size == 0) {
+                        DomainVerificationResult(
+                            ontologySuccess = true,
+                            useCaseSuccess = false,
+                            "No workflows found"
+                        )
+                    } else {
+                        DomainVerificationResult(ontologySuccess = true, useCaseSuccess = true, null)
+                    }
+                }
+
+                domainOperation.saveVerification(DomainVerification(apeRequest.domain.id, result))
+                return result
+            } catch (exc: Exception) {
+                val result = when (exc) {
+                    is FileNotFoundException ->
+                        DomainVerificationResult(current.ontologySuccess, null, exc.message)
+                    is SynthesisFlagException, is RunParametersExceedLimitsException ->
+                        DomainVerificationResult(current.ontologySuccess, false, exc.message)
+                    else ->
+                        throw exc
+                }
+                domainOperation.saveVerification(DomainVerification(apeRequest.domain.id, result))
+                return result
+            }
+        } catch (exc: Exception) {
+            when (exc) {
+                is AccessDeniedException ->
+                    throw ResponseStatusException(HttpStatus.FORBIDDEN, exc.message, exc)
+                else ->
+                    throw ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "", exc)
+            }
+        }
+    }
+
+    /**
      * The user can send a configuration and is then returned as a zip containing the configuration
      * @param inputData
-     * @return an response entity containing the Response entity with the resource
+     * @return a response entity containing the Response entity with the resource
      */
     @ResponseStatus(HttpStatus.OK)
     @PostMapping("/config")