+++ back-end/src/main/kotlin/com/apexdevs/backend/web/controller/entity/domain/DomainDetails.kt
@@ -7,7 +7,18 @@ package com.apexdevs.backend.web.controller.entity.domain
 import com.apexdevs.backend.persistence.database.entity.DomainVisibility
 
 /**
- * Used to send domain information to FE.
+ * Used to send more detailed domain information of a specific domain to FE.
+ *
+ * @param id the unique id of the domain
+ * @param title name of the domain
+ * @param topics list of the topics the domain belongs to
+ * @param description the description of the domain
+ * @param visibility The visibility level of the domain.
+ * @param topics The topics related to the domain.
+ * @param ontologyPrefixIRI The ontology prefix IRI.
+ * @param toolsTaxonomyRoot The tools' taxonomy root.
+ * @param dataDimensionsTaxonomyRoots The data dimensions' root.
+ * @param strictToolAnnotations Whether the domain uses strict tool annotations.
  */
 data class DomainDetails(
     val id: String,
@@ -18,5 +29,5 @@ data class DomainDetails(
     val ontologyPrefixIRI: String,
     val toolsTaxonomyRoot: String,
     val dataDimensionsTaxonomyRoots: List<String>,
-    val strictToolsAnnotations: Boolean
+    val strictToolAnnotations: Boolean
 )