+++ front-end/src/models/Domain.tsx
@@ -7,6 +7,9 @@
 
 /**
  * An APE domain
+ *
+ * @deprecated Has been replaced by the more specific
+ * {@link DomainInfo}, {@link DomainDetails} and {@link DomainWithAccess} types.
  */
 export default interface Domain {
   /** The ID of the domain */
@@ -51,6 +54,58 @@ export interface DomainInfo {
   visibility?: Visibility,
   /** Optional access level for the current user. */
   access?: Access,
+  /** Domain verification status. */
+  verification?: DomainVerificationResult,
+}
+
+/**
+ * Domain information with more details when requesting a specific domain from the back-end.
+ *
+ * Used when exploring or editing a domain.
+ */
+export interface DomainDetails {
+  /** The ID of the domain. */
+  id: string,
+  /** The name of the domain. */
+  title: string,
+  /** The description of the domain. */
+  description: string,
+  /** The visibility level of the domain. */
+  visibility: Visibility,
+  /** The topics related to the domain. */
+  topics: string[],
+  /** The ontology prefix IRI. */
+  ontologyPrefixIRI: string,
+  /** The tools taxonomy root. */
+  toolsTaxonomyRoot: string,
+  /** The data dimensions taxonomy root. */
+  dataDimensionsTaxonomyRoot: string[],
+  /** Whether the domain uses strict tool annotations. */
+  strictToolAnnotations: boolean,
+}
+
+/**
+ * Domain with additional access information.
+ *
+ * Used when fetching all domains a certain user has access to.
+ */
+export interface DomainWithAccess {
+  /** The ID of the domain. */
+  id: string,
+  /** The name of the domain. */
+  title: string,
+  /** The description of the domain. */
+  description: string,
+  /** The visibility level of the domain. */
+  visibility: Visibility,
+  /** The topics related to the domain. */
+  topics: string[],
+  /** The ID of the user who's access this is about. */
+  userId: string,
+  /** Access level for the current user. */
+  access: Access,
+  /** Domain verification status. */
+  verification: DomainVerificationResult,
 }
 
 /**
@@ -111,4 +166,16 @@ export interface UserAccessUpload {
   userId: string,
   /** The access level the user will get to the domain. */
   access: Access,
+}
+
+/**
+ * Object received from the back-end, whether a domain verification was successful.
+ */
+export interface DomainVerificationResult {
+  /** Whether the ontology was successfully verified. */
+  ontologySuccess?: boolean,
+  /** Whether the use case configuration was successfully verified. */
+  useCaseSuccess?: boolean,
+  /** A description why the verification failed, when applicable. */
+  errorMessage?: string,
 }