+++ CHANGELOG.md
@@ -6,6 +6,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.6.0] - 2022-04-22
+
+### Added
+
+- Domain verification. Domains' ontology and core configuration are verified with an empty run, and use cases are verified to run and give at least one workflow result.
+
+### Fixed
+
+- Buttons "Go back home" and "Go to domain" after creating a domain are now properly centered.
+
 ## [1.5.1] - 2021-12-24
 
 ### Changed
@@ -18,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Clear buttons to clear all input data, output data, constraints on the explore page.
 - Domains can now be permanently deleted in the domain edit page.
+- Domains' ontologies and use cases are now verified when created.
 
 ### Changed
 
@@ -181,6 +192,7 @@ This release is paired with the APE Web [DOI release](https://zenodo.org/badge/l
   * Approve user accounts.
   
 [Unreleased]: https://github.com/sanctuuary/APE-Web/compare/master...dev
+[1.6.0]: https://github.com/sanctuuary/APE-Web/compare/v1.5.1...v1.6.0
 [1.5.1]: https://github.com/sanctuuary/APE-Web/compare/v1.5.0...v1.5.1
 [1.5.0]: https://github.com/sanctuuary/APE-Web/compare/v1.4.0...v1.5.0
 [1.4.0]: https://github.com/sanctuuary/APE-Web/compare/v1.3.3...v1.4.0